+++ src/lib/DebouncedBooleanInterface.java
@@ -1,0 +1,6 @@
+package lib;
+
+public interface DebouncedBooleanInterface {
+    void add(boolean b);
+    boolean get();
+}
